made header and modified category model

+++ app/Http/Controllers/Admin/ControllerSetting.php
@@ -16,6 +16,10 @@ class ControllerSetting extends Controller
         $data['error'] = false;
         if (isset($request->error)) $data['error'] = $request->error;
 
+        $data['address'] = Setting::where('version', '=', 'address')->get();
+        $data['socials'] = Setting::where('version', '=', 'socials')->get();
+        $data['phones'] = Setting::where('version', '=', 'phones')->get();
+
         $data['tab_1'] = Setting::where('tab', 1)->get();
         $data['tab_2'] = Setting::where('tab', 2)->get();
         $data['tab_3'] = Setting::where('tab', 3)->get();
@@ -25,16 +29,20 @@ class ControllerSetting extends Controller
     }
 
     public function update(Request $request){
-
+        //dd($request->all());
         //Проверка на уникальность полей
         $error = false;
         $arr = [];
         $items = $request->all();
-        foreach ($items['name'] as $key=>$item) {
-            if (isset($arr[$items['name'][$key]])) {
-                $arr[$items['name'][$key]]++;
-            }else {
-                $arr[$items['name'][$key]] = 1;
+       
+        if ( isset($items['name']) ) {
+            foreach ($items['name'] as $key=>$item) {
+                if (isset($arr[$items['name'][$key]])) {
+                    $arr[$items['name'][$key]]++;
+
+                }else {
+                    $arr[$items['name'][$key]] = 1;
+                }
             }
         }
 
@@ -51,19 +59,22 @@ class ControllerSetting extends Controller
         if ($error) {
             return redirect()->route('settings.index', $error);
         }else {
+            if ( isset($items['name']) ) {
 
-            Setting::truncate();
-            foreach ($items['name'] as $key=>$item) {
-                $setting = new Setting();
+                Setting::truncate();
+                foreach ($items['name'] as $key=>$item) {
+                    $setting = new Setting();
+    
+                    $setting->name = $items['name'][$key];
+                    $setting->value = $items['value'][$key];
+                    $setting->descr = $items['descr'][$key];
+                    $setting->type = $items['type'][$key];
+                    $setting->tab = $items['tab'][$key];
+                    $setting->version = $items['version'][$key];
 
-                $setting->name = $items['name'][$key];
-                $setting->value = $items['value'][$key];
-                $setting->descr = $items['descr'][$key];
-                $setting->type = $items['type'][$key];
-                $setting->tab = $items['tab'][$key];
-
-                $setting->save();
-
+                    $setting->save();
+    
+                }
             }
 
             return redirect()->route('settings.index');
